[refactor] SportsServiceList 파일들 패키지 이동
